src/test/java/cn/vitalking/service/impl/ProductCategoryServiceImplTest.java
package cn.vitalking.service.impl;

import cn.vitalking.entity.ProductCategory;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.Arrays;
import java.util.List;

@RunWith(SpringRunner.class)
@SpringBootTest
public class ProductCategoryServiceImplTest {

    @Autowired
    private ProductCategoryServiceImpl productCategoryService;

    @Test
    public void findById() throws Exception {
        ProductCategory productCategory = productCategoryService.findById(2);
        System.out.println("productCategory:" + productCategory);
        Assert.assertEquals(new Integer(2), productCategory.getCategoryId());
    }

    @Test
    public void findAll() throws Exception {
        List<ProductCategory> list = productCategoryService.findAll();
        Assert.assertNotEquals(0, list.size());
    }

    @Test
    public void findByCategoryTypeIn() throws Exception {

        List<ProductCategory> list = productCategoryService.findByCategoryTypeIn(Arrays.asList(1, 2, 3, 4));
        Assert.assertNotEquals(0, list.size());
    }

    @Test
    public void save() throws Exception {
        ProductCategory  productCategory = new ProductCategory("社会科学",7);
        ProductCategory result = productCategoryService.save(productCategory);
        Assert.assertNotEquals(null, result);
    }

}